prototype for BilinearForm interface
This new API can be used to assemble a DPG system with arbitrary number
of trial and test spaces.
Bilinear form and test space inner product can be given as sums of
integral terms.

For an example of how to use this new API, we refer to the transport
problem in src/dune_dpg.cc.

Important new classes:
* SystemAssembler:
    This new class handles the assembly of the system matrix and rhs
    vector.
* BilinearForm:
    Handles the computation of the local matrices and the occupation
    pattern corresponding to a bilinear form.
* InnerProduct:
    Handles the computation of the local matrices and the occupation
    pattern corresponding to an inner product.