IntegralTerm: only compute what we use
This puts the evaluation of test functions or their gradients into
seperate functions and IntegralTerm::getLocalMatrix() now only calls
the evaluation functions it really needs.

There is still room for improvements in the case that we have the same
space twice with the same kind of evaluations, but that would make the
code more complicated again.