interpolate DOFs on characteristic faces
Since the DOFs on the interior of characteristic faces are not well
defined, we add additional constraints to the system matrix that
interpolate these DOFs from the endpoints of the characteristic face.
Since the endpoints have to belong to non-characteristic faces, they
are already well-defined.
Thus our interpolation makes the matrix non-singular.