Swapped defineCharacteristicFaces and applyDirichletBoundarySolution.
The definition of coefficients in the matrix to impose boundary conditions (done in applyDirichletBoundarySolution) was creating a conflict with the interpolation of degrees of freedom that are on characteristic faces (applyDirichletBoundarySolution). The problem is that some coefficients for the interpolation were set to zero. This was solved by changng the order of the two methods.