fix regression with boost fusion 1.58
The fold operation in boost fusion 1.58 seems to be more strict when it
comes to the types of its arguments. Before, fold would properly
propagate the types, when supplied with a list of size_t and an int as
initial state. Also it was a bit less strict with the types of the
arguments of its function-argument.

Now we make sure to only feed size_t values to fold and we changed the
first argument of offsetHelper and globalOffsetHelper to const size_t&
instead of size_t.