introduce LinearForm
This is a new abstraction used for linear forms like they appear in the
rhs. That should make it easier to later add additional terms given by
non-zero boundary conditions.

We also split applyDirichletBoundary into matrix and rhs contributions.
This is only implemented to set boundary conditions for the solution
space for now.