add license and copyright statements
After discussing with Angela, Olga and Valentina, we agreed to license
our code under GPL2 with runtime exception to be compatible to the
license of the dune core libraries.

Since the dune core libraries might switch to LGPL3 in the near future,
we will probably dual-license dune-dpg under LGPL3 or later. But first
I would like to wait for any discussions on licensing on the dune
mailing lists.