make_DPG_SystemAssembler doesn't need innerProduct
The SystemAssembler for a DPG system defines innerProduct to a nullptr
now, as only the innerProduct of the optimal test space is needed.

Thus make_DPG_SystemAssembler does not take an innerProduct argument
anymore.

To make this work, we had to factor out some code specific to the
saddlepoint formulation that was accessing innerProduct.