[cleanup] use auto to remove forward-declaration
The forward-declaration of UGGrid<dim> can be circumvented by using the
auto keyword. This makes the code more robust to changes in the
refinedReferenceElement() method.
This forward-declaration was also quite confusing as it had to be done
at a place where it was not directly obvious why it was needed at all.