now the scattering assemblers share a common SVD
This way we have to compute the SVD of the scattering kernel matrix only
once.

We also renamed some classes and functions to better reflect what they
implement.

This change implicitly defines an interface that can be used to define
alternative approximations of the scattering kernel. Time will tell
whether this interface fits the needs of other approximations.