[bugfix] fix refined finite elements
When writing the integral term code for refined finite elements, I
forgot to change gridView into referenceGridView at one place.
Thus, I computed the intersections from the global gridView and an
element from the reference grid. This obviously leads to undefined
behaviour, but wasn't caught as the types matched and UG didn't complain
during runtime that the given element didn't belong to the given
gridView.
We caught this bug more by chance, when trying to use AlbertaGrid for
the global grid, as the the types of gridView and element wouldn't fit
then anymore.