[bugfix] fix ambiguity in tuple construction
GCC 6 throws an error, because the call to the constructor of
std::tuple in `make_LinearIntegralTerm` was ambiguous.

It seems GCC 5 didn’t know about the constructor

```
    template< class Alloc >
    tuple( std::allocator_arg_t, const Alloc& a );
```

By replacing the direct call to the constructor of std::tuple
with std::make_tuple, we remove this ambiguity and also don’t need
to specify the the tuple types twice.