new Geometry buffer for DPGSystemAssembler
Use only for constant coefficients!

The geometry and the coefficient matrix for the optimal test space
as well as the corresponding local matrix of the bilinear form are
saved in an std::map and thus can be reused on the same
(but possibly shifted) triangle. This reduces the time for assembling
the matrix. (The rate of reduction depends on the number of occuring
different triangles)

Note that a similar buffer could be helpful in errortools
for the computation of the aposteriori error.

plot_solution modified as test example for geometry buffer

added measurements of runtime of different parts in plot_solution