bind testspaceCoefficientMatrix earlier and removed auxillary variables
In assembleSystem and in assembleRhs, the testspaceCoefficientMatrix
now is bound before rhsLinearForm.getLocalVector(...) is called.
This way, somehow a mysterious segementation fault in Convergence test does
not occure anymore, although I have no idea why. Note that this probably
does not really fix the bug, it is only a dirty solution which seems to
work right now but we still have to find that bug.