[Periter] handle boundary values
The boundary values are given to Periter as a function for the lifting g
of the boundary values and a function for s·∇g.
Since we have to project g to the finite element space of the inner DoFs
anyway, it might be more wise to only take g and compute a FE
approximation from it from which we then compute the derivative.
The error we make by this approach would then be part of the rhs
approximation error.

For now, the boundary value is still set to a constant 0 function.