update INSTALL instructions to Dune 2.5
Mention correct URL for the Dune 2.5 release. Dune-typetree,
dune-functions and dune-uggrid are all part of the 2.5 release, so we
don't have to tell the user to get them from Git anymore.
We have also removed the deprecated instructions for installing UG which
has been superseeded by the more user-friendly dune-uggrid.