Periter: use dune-subgrid and HangingNodeP2NodalBasis
We need to depend on dune-subgrid >= 2.6, as previous versions had a bug
in the intersection iterator which caused it to skip over intersections
in certain situations.

We switch the for loops in periter, so that in the inner iteration, we
first iterate over all directions and then successively refine for each
direction until the required accuracy is reached.

We now use adaptive grid refinements to get one well-adapted grid for
each transport direction. The computation of the scattering integral
still has to be done on the host grid into which the sub grids with the
transport solutions are embedded. To this end, we use the previously
defined interpolation and projection routines for prolongation and
restriction, respectively.