HangingNodeP2NodalBasis
This class handles hanging node constraints for 2nd degree Lagrangean
finite elements on triangular grids. This uses the
ConstrainedGlobalBasis interface to provide an index set that is capable
of mapping to constrained global indices. The guiding ideas for
implementing this index set are taken from Kronbichler and Kormann, 2012
(doi: 10.1016/j.compfluid.2012.04.012).