generic iteration over index sets
The newly introduced function iterateOverLocalIndexSet abstracts the
iteration over an index set. It can be used to copy from a local vector,
living only on one element, to a global vector. The user has to provide
functions that describe the copy operation for a given local and global
index and one function describing the copy operaton for a constrained
degree of freedom (taking a local index, a global index and a weight).