workarounds for dune-subgrid
dune-subgrid does not implement the whole Grid interface, in particular
the SubGrid::Intersection class does not implement the
centerUnitOuterNormal(), geometryInInside() and conforming() methods.

We provide wrapper functions for these methods that for normal Grids
simply forward to the respective method of the Intersection while
providing an ersatz implementation for SubGrid::Intersection.

squash! workarounds for dune-subgrid