projection from HostGrid to SubGrid
The function attachDataToSubGrid creates an object that holds the data
from the host grid and the projection to the current sub grid in a way
that we only need to project to newly introduced cells after further
refinements.

The projection to the sub grid will be created by the method
restoreDataToRefinedSubGrid. It iterates over all cells and projects
only where necessary and interpolates when we refine beyond the initial
host grid.