[cleanup] use Element::*my*dimension
Since we use this dimension to specify the type of quadrature points,
mydimension is the right dimension. In our case, we are always on
codimension 0 elements, i.e. mydimension==dimension. This change should
however make our intentions a little bit more clear.

This is a follow-up to commit 652f813761045567582a7c54cc5dd34725e4238a.