[cleanup] remove SubGridGlobalBasisRestrictor
This class was ment to implement a restriction to a subgrid using the
`transfer` method of the `SubGrid` class. The interface of the
`transfer` method makes it near impossible to implement a restriction
by projection. Thus `SubGridGlobalBasisRestrictor` was never started
to be implemented.

If you want to restrict a host grid function to the subgrid, use the
functions in `/dune/dpg/radiative_transfer/subgridprojection.hh`
instead. These are the `attachDataToSubGrid` function and the
`restoreDataToRefinedSubGrid` method of the returned
`SubGridProjectionData` object.