[cleanup] don't hold reference to temporary
In this particular case the reference to temporary would still be valid
as the const reference would prolong the lifetime of the object it
references to. However directly initializing an object instead of
creating a temporary and holding a reference to it makes it more obvious
that the object is still valid.