[bugfix] make temporary copy of local data
Somehow giving RestoreDataToRefinedGridFunction an iterator into
localData resulted in wrong code when compiling in Release mode. Copying
the required subrange of localData into a temporary vector and giving
that to RestoreDataToRefinedGridFunction prevents the problem with
generating wrong code.

I still don't understand why the code using an iterator didn't work. Is
my understanding of iterators wrong or do I have another bug in the
code or is this even a compiler error?