Periter: use intersection of neighboring grids
When refining in the angular discretization, we have to introduce new
grids for the newly added transport directions. Before we just copied a
neighboring grid, but that was already highly refined in the wrong
direction. To keep the number of DoFs as low as possible we now take the
intersection of the two neighboring grids to obtain a starting grid that
is coarse enough to not introduce too many unnecessary DoFs while still
being roughly refined around the right places.