Periter: use harmonic extension of boundary values
To handle non-zero boundary values, we need a lifting of them. Because
computing a seperate lifting for each transport direction would be
costly, we take the same harmonic extension of the boundary values for
all transport directions. The harmonic extension has the benefit of
being smooth enough to be in all the directed Sobolev spaces associated
to the transport directions.

To make our computations simpler we now use alternative 2 of [BDS],
Remark 3.6 to handle the boundary values.

Broersen, Dahmen and Stevenson give two ways of handling non-zero
boundary values. Until now we used the first alternative which needed
the derivative of the lifted boundary values. The second alternative
only needs the lifting itself and thus should be simpler to use when we
now decouple the boundary values from the rest of the rhs.