new SkeletalLinearFunctionalTerm
This new class describes a term in a LinearForm that integrates test
functions against a fixed function given in a finite element space on
the skeleton of the grid.
One application for it would be in the second alternative for handling
boundary values of Broersen, Dahmen, Stevenson Remark 3.6. There an
integral over the skeleton (of the refined grid on which the test
functions are defined) is used in the rhs of the formulation.