[bugfix] make computation of splitPoint more robust
As it turned out commit 1894cdb0a50705655ee19391bee95894468f7a6f did not
fix the bug in the computation of splitPoint. The assertion just
triggered later during refinement.

What seems to be the problem are numerical inaccuracies that lead
referenceBeta to be almost 0 in one coordinate. But since the
transformations from beta to referenceBeta depend on the scale of the
element on which we are these numerical inaccuracies become larger
during refinement.
To fix this problem we now look at a relative error w.r.t. the norm of
referenceBeta when comparing components to 0. That should counter the
scale dependence of the numerical inaccuracies.