script for visualizing solution and transport direction
This script takes a series of pictures generated with Visit or ParaView
and the output file generated by dune_dpg_rad_trans_periter and
generates a video of the solutions with corresponding transport
directions.

The pictures of the solution have to be named u-n*-dir0000.png or
u-elevate-n*-dir0000.png if you use the --elevate option.