fix error estimate and definition of rhobar
There was a mistake in our paper in the error computation that could
lead to nonconvergent iterations. This new error error estimate does not
suffer from the same problem, but it sets a significantly stricter
accuracy. This requirement could be further reduced, e.g. by setting
eta_n = rho/(1+n^2).

Thanks to Olga for correcting the error estimates.