don't use a temporary GridView in elements()
The `elements` function returns an `IteratorRange` object that only
saves a copy of the begin() and end() iterators of the GridView. The
GridView itself is not copied into the IteratorRange. Thus the iterators
in the IteratorRange might be invalidated if we provide the `elements`
function with a temporary gridView.

To prevent any troubles, make sure that the `elements` function is only
called on an lvalue (named variable) GridView.