Periter: start transport solve with merged grid
Previously, we used an intersection of neighboring subgrids to get a
smaller grid for the transport solve. This did however pose the problem
that we projected the extension of boundary values to these
intersections of subgrids which could destroy the continuity of the
extension.
When starting with the merged grid, we know that the boundary extension
will always be in H_s.