[cleanup] remove old SVD kernel approximation
This has been superseded by the wavelet-based SVD kernel approximations
that have the benefit that the nestedness of wavelet spaces easily
allows to increase the number of directions during runtime.

The old SVD kernel approximation wasn't usable since a long time ago
when the interfaces in the kernel approximations changed. Since we will
not need it anymore, it's better to remove this cruft.