[bugfix] prevent compiler from evaluating wrong eval_if branch
With Clang we got an error in an eval_if branch for refined finite
elements when we were actually using it with unrefined ones. We now make
this eval_if branch dependent on the provided identity function argument
so that it can only get analyzed by the compiler if we use a refined
finite element.

Somehow GCC didn't look at that eval_if branch even before this change.
Seems that Clang is doing a stricter analysis here.