Periter: allow different ratios between kappas
The parameters \kappa_{1,2,3}, which describe how the error is divided
between scattering, rhs approximation and transport solver (in that
particular order), have previously computed in a way that equilibrates
those errors.

Now the ApproximationParamters constructor takes a value accuracyRatio
that says how the errors are split up between scattering and transport
solver.

accuracyRatio is a value between 0 and 1 where closer to 1 means more
accuracy for the transport solver and closer to 0 means more accuracy
for the kernel approximation.