fix periter_plot.py with newer matplotlib versions
Some months ago an update of the matplotlib package broke the generation
of plots on my Debian machine. It seems that matplotlib changed the way
in which the plot command interprets input strings.

We now convert all input values to int or float after parsing to fix
this problem. This also seems to make the plot script a little bit
faster as the plot functions now don't have to convert their input
anymore.