[bugfix] workaround for wrongly reconstructed SubGrids
When calling restoreSubGridFromIdSet, we sometimes obtained a SubGrid
that contained cells that were not in the given idSet. It seems that
this has to do with the max level difference which leads SubGrid to
wrongly refine a cell.

Some tests on a small example case showed that these wrongly refined
cells can be prevented by using `insertSetPartial` instead of
`insertSet`.