subgridprojection.hh: remove restriction in interpolation
Previously, when interpolating to a subgridcell, we had the restriction
that this subgridcell had to be a direct descendant of the cell from
which we interpolate.
This restriction was only due to lazyness, so that we could use the
child.geometryInFather(). It has now been replaced with a more general
geometry that works with arbitrary levels of descendancy.