simplify for loop
By directly continuing when child is a leaf, it is obvious that we only
process leaf elements. Previously you had to scroll almost a whole
screen full of code, to the end of the if body to see that there was no
else case and nothing after the if statement.