handle case of mixed projection/interpolation
When refining cells twice (e.g. to reinstate level 1 difference between
neighbors), we could get the situation that some subgrid cells, that were
previously below the leaf level of the original host grid, have been
(partially) refined beyond the leaf level of the original host grid.

Thus we have to handle interpolation in addition to the projection that
we were always doing in the case of cellData that has associated data
from descendants.