Periter: change the way we compute the kernel accuracy
Previously, we wrongfully used an estimate from best N-term
approximation for our a priori estimates on the required wavelet level.
This estimation was not correct as we confused a sparseness parameter
with the wavelet level.

What we do now is that we compute a large truth matrix, whose wavelet
level we choose sufficiently large (by taking a conservative estimate
that is comparable with the number of transport directions normally used
in the radiative transfer literature). We assume that this truth matrix
has a negligible truncation error when compared to the exact wavelet
representation of the scattering kernel.
Then we can compute truncation errors for smaller wavelet levels by
computing the error between those smaller matrices and the truth matrix.

For the SVD approximation we make sure that the sum of truncation error
and error of low-rank approximation is below the required scattering
accuracy.