Periter: remove outer iteration 0 from plots
The 0th outer iteration is a special case where the rhs is still
initialized to 0. Thus we can compute the scattering exactly for this
iteration step and get scattering error = 0. This produces a bump at the
beginning of the iteration, when the scattering error jumps from 0 to a
proper estimate in outer iteration 1.

To suppress this distracting bump, we simply remove the setup phase of
outer iteration 0 in our convergence plots.