Revert "remove unnecessary string conversions"
Somehow this commit slightly changes the behaviour of Periter.
When called as

export KAPPA_RATIO=0.2
./periter_checkerboard_normalized -p -i \
  -o "../results/checkerboard/normalized/$KAPPA_RATIO/" \
  -k $KAPPA_RATIO -n 64 0.005 0.5 11 7

I get small differences in the error estimates in iteration 9.1,
direction 15, that lead to a cascade of differences due to different
marking of cells for refinement.

I do not see how changing those strings could have introduced those
differences and I cannot say whether old or the new results are
"better". So for now I revert this commit before this phenomenon is more
thoroughly investigated.

This reverts commit 63da342654266fa991c580cb6f21881b450265f7.