fix subgrid projection for unrefined spaces
While `findSubElementContainingChild` is not used for unrefined spaces,
we still got a compiletime error, as we were asking for the
RefinementGridView of the space in the return type of this function.
By introducing a new template parameter and asking this new template
parameter for the RefinementGridView, we can make sure that the typename
involving RefinementGridView is not evaluated for unrefined spaces.