DoerflerMarking: remove potential source of nondeterminism
In the `DoerflerMarking` function, we have to sort the ElementSeeds by
descending error estimate. This could lead to a non-deterministic
marking of cells when there are two elements with the same error
estimate just before and after the cut-off point.

By using a stable sort algorithm, we remove this source of
non-determinism, as now, after sorting, elements with the same error
estimate are given in the order provided by the GridView.