Reduce number of memory allocations.
This reduces the number of memory allocations for local function evalutations.

This reduced the runtime of the asti_checkerboard_normalized example
with 11 iterations and parameters chosen as in my thesis from
14h 45min to 9h59min on the cluster of TU/e's math department.
These are more than 30% reduction of the runtime!

The results started to slightly deviate in iteration 9. Probably because
of slightly different rounding that then might have lead to slightly
different grid refinement. The differences in the error estimators and
number of degrees of freedom was however minor and visually there was no
difference in the plots of the solution in the last iteration.