don't resize localVector in copyToLocalVector
First of all, it is misleading that a function called copy does a
resize. Second, the resize is redundant in all but one calls of this
function in our codebase. And third, it doesn't work when passing an
iterator as the target of the copy operation. This means that the code
in refinementinterpolation.hh did not work after commit
b74728f3f3ce71c1e892a3aabc3f800de1741c53.