Prevent unnecessary instantiation of NormalizedPreBasis.
In NormalizedPreBasis::makeNode(), I erroneously instantiated the
NormalizedNode from innerProduct_ instead of *this. This worked by
accident because innerProduct_ was implicitly converted to a new
NormalizedPreBasis that was then given to the constructor of
NormalizedNode.

I'm now initializing the Node directly from the PreBasis. I've also
marked the constructors of NormalizedPreBasis and NormalizedNode as
explicit to prevent such unintended implicit conversions.