Define derivative of piecewise constant GV function.
This defines the derivative of PiecewiseConstantGridViewFunction as a
constant-0 function.

While we don't use the derivative directly, it has become necessary to
define it to work around the fact that dune-function's interpolate
function now requires the derivate to be defined since they recently
implemented Hermite interpolation. It was probably unintended on their
side to require derivatives even if one only uses Lagrange
interpolation.